requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
sqlalchemy>=2.0,<3
httpx>=0.27,<1
pydantic-settings>=2.6,<3
openai-agents>=0.1.0
python-dotenv>=1.0,<2
python-multipart>=0.0.9,<1
pytest>=8.3,<9
pytest-asyncio>=0.24,<1
